release resources better between tests, export a library that modules like apostrophe-site-map can use to do cleanup of apostrophe between tests
