+++ lib/modules/apostrophe-attachments/index.js
@@ -210,6 +210,14 @@ module.exports = {
       }
     };
 
+    self.apostropheDestroy = function(callback) {
+      if (!self.uploadfs.destroy) {
+        self.apos.utils.warn('uploadfs dependency is old, cannot clean up resources i.e. timers that will prevent exit');
+        return callback(null);
+      }
+      return self.uploadfs.destroy(callback);
+    };
+
     self.enableCollection = function(callback) {
       return async.series([ collection, indexDocIds, indexTrashDocIds ], callback);
       function collection(callback) {
